Begin simplifying the test utils
