+++ examples/async-graphql/tests/test_user_integration.rs
@@ -40,7 +40,7 @@ async fn test_user_get_current_simple() -> Result<()> {
     let utils = TestUtils::init().await?;
 
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     // Create a user with this username
     let users = utils.app.get(&USERS_SERVICE).await?;
@@ -86,7 +86,7 @@ async fn test_user_get_current_no_user() -> Result<()> {
     let utils = TestUtils::init().await?;
 
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     let req = utils
         .graphql
@@ -131,7 +131,7 @@ async fn test_user_get_or_create_current_existing() -> Result<()> {
     let utils = TestUtils::init().await?;
 
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     // Create a user
     let users = utils.app.get(&USERS_SERVICE).await?;
@@ -179,7 +179,7 @@ async fn test_user_get_or_create_current_create() -> Result<()> {
 
     let email: String = FreeEmail().fake();
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     let req = utils.graphql.query(
         GET_OR_CREATE_CURRENT_USER,
@@ -264,7 +264,7 @@ async fn test_user_update_current() -> Result<()> {
     let utils = TestUtils::init().await?;
 
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     // Create a user with this username
     let users = utils.app.get(&USERS_SERVICE).await?;
@@ -342,7 +342,7 @@ async fn test_user_update_current_requires_user() -> Result<()> {
     let utils = TestUtils::init().await?;
 
     let username = Ulid::new().to_string();
-    let token = utils.create_jwt(&username);
+    let token = utils.create_jwt(&username).await?;
 
     let req = utils.graphql.query(
         UPDATE_CURRENT_USER,
