+++ nakago_async_graphql/src/test/mod.rs
@@ -1,0 +1,2 @@
+/// HTTP testing utils
+pub mod http;
