+++ nakago_axum/src/test/mod.rs
@@ -1,0 +1,4 @@
+/// HTTP testing utilities
+pub mod http;
+
+pub use http::HttpClientProvider;
