+++ website/tsconfig.json
@@ -1,6 +1,7 @@
 {
     "extends": "@tsconfig/docusaurus/tsconfig.json",
     "compilerOptions": {
-        "baseUrl": "."
+        "baseUrl": ".",
+        "module": "Node16"
     }
 }
